requirements.txt
torchmetric
jiwer
pandas
openpyxl
text_normalize
streamlit
num2words
